+++ apps/js-sdk/firecrawl/src/v2/methods/batch.ts
@@ -8,6 +8,7 @@ import {
 } from "../types";
 import { HttpClient } from "../utils/httpClient";
 import { ensureValidScrapeOptions } from "../utils/validation";
+import { fetchAllPages } from "../utils/pagination";
 import { normalizeAxiosError, throwForBadResponse } from "../utils/errorHandler";
 
 export async function startBatchScrape(
@@ -71,7 +72,7 @@ export async function getBatchScrapeStatus(
       };
     }
 
-    const aggregated = await fetchAllBatchPages(http, body.next, initialDocs, pagination);
+    const aggregated = await fetchAllPages(http, body.next, initialDocs, pagination);
     return {
       status: body.status,
       completed: body.completed ?? 0,
@@ -136,42 +137,3 @@ export function chunkUrls(urls: string[], chunkSize = 100): string[][] {
   for (let i = 0; i < urls.length; i += chunkSize) chunks.push(urls.slice(i, i + chunkSize));
   return chunks;
 }
-
-async function fetchAllBatchPages(
-  http: HttpClient,
-  nextUrl: string,
-  initial: Document[],
-  pagination?: PaginationConfig
-): Promise<Document[]> {
-  const docs = initial.slice();
-  let current: string | null = nextUrl;
-  let pageCount = 0;
-  const maxPages = pagination?.maxPages ?? undefined;
-  const maxResults = pagination?.maxResults ?? undefined;
-  const maxWaitTime = pagination?.maxWaitTime ?? undefined;
-  const started = Date.now();
-
-  while (current) {
-    if (maxPages != null && pageCount >= maxPages) break;
-    if (maxWaitTime != null && (Date.now() - started) / 1000 > maxWaitTime) break;
-
-    let payload: { success: boolean; next?: string | null; data?: Document[] } | null = null;
-    try {
-      const res = await http.get<{ success: boolean; next?: string | null; data?: Document[] }>(current);
-      payload = res.data;
-    } catch (err: any) {
-      // axios throws on non-2xx; break pagination loop on fetch error
-      break;
-    }
-    if (!payload?.success) break;
-    for (const d of payload.data || []) {
-      if (maxResults != null && docs.length >= maxResults) break;
-      docs.push(d as Document);
-    }
-    if (maxResults != null && docs.length >= maxResults) break;
-    current = (payload.next ?? null) as string | null;
-    pageCount += 1;
-  }
-  return docs;
-}
-
